[quant][graphmode] Rename _quantize_script.py to quantize_script.py
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
