+++ mypy.ini
@@ -67,7 +67,7 @@ ignore_errors = True
 [mypy-torch.quantization.fake_quantize]
 ignore_errors = True
 
-[mypy-torch.quantization._quantize_script]
+[mypy-torch.quantization.quantize_script]
 ignore_errors = True
 
 [mypy-torch.quantization._numeric_suite]
